roles/slg.galaxy_stats/files/requirements.txt
influxdb
psycopg2-binary
pyyaml
pycryptodome
python-dotenv
sqlalchemy
